requirements.txt
discord.py >= 1.3.3
psycopg2-binary >= 2.8.5
pandas >= 1.0.5
asyncio >= 3.4.3
aiohttp >= 3.6.2
#asyncpg >= 0.20.1